refactor front: move task input creation to separate file
